Script 'remodeling' -- Still in progress
I am close to at least making the Script more readable -- and hopefully improvable. Moving forward, it may be easier for me to start afresh and keep this for reference. In this Script a 'caching' system has been more the focus, to the unfortunate detriment of the quality (i.e. beauty) of the tree generator; and I hope that in subsequent editions, this will not be the case. Furthermore, I should like to shift from the classic "pseudo-random tree generation" style toward more of a "procedural tree simulation" style, distinct from the former in that the focus would be to 'simulate' tree growth, accounting for factors contributing toward or against its growth, such as water supply, available sunlight, nutrition, health, competition with other plants, root growth, and obstacle avoidance. Such an endeavor is, as far as I can tell, a lot more difficult of a task than to do what I have done thus far (in the past, making pseudo-random tree generators). More thought and grace is required.